requirements.txt
fastapi==0.115.13
PyJWT==2.10.1
python-multipart==0.0.20
PyYAML==6.0.2
uvicorn==0.34.3
